Add casts for Windows

+++ src/mono/mono/eglib/giconv.c
@@ -305,7 +305,7 @@ g_utf16_to_utf8_impl (const gunichar2 *str, glong len, glong *items_read, glong 
 		flags |= MINIPAL_TREAT_AS_LITTLE_ENDIAN;
 #endif
 	if (len < 0) {
-		len = minipal_u16_strlen (str) + 1;
+		len = (glong)minipal_u16_strlen (str) + 1;
 	}
 
 	glong ret = (glong)minipal_get_length_utf16_to_utf8 (str, len, flags);
@@ -346,7 +346,7 @@ g_utf16_to_utf8_custom_alloc (const gunichar2 *str, glong len, glong *items_read
 	errno = 0;
 
 	if (len < 0) {
-		len = minipal_u16_strlen (str) + 1;
+		len = (glong)minipal_u16_strlen (str) + 1;
 	}
 
 	glong ret = (glong)minipal_get_length_utf16_to_utf8 (str, len, 0);
@@ -384,7 +384,7 @@ g_utf16_to_ucs4 (const gunichar2 *str, glong len, glong *items_read, glong *item
 	g_return_val_if_fail (str != NULL, NULL);
 
 	if (len < 0) {
-		len = minipal_u16_strlen (str);
+		len = (glong)minipal_u16_strlen (str);
 	}
 
 	inptr = (char *) str;
